backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.11.7
pydantic-settings==2.4.0
python-multipart==0.0.9
google-generativeai==0.8.3
httpx==0.27.0
tenacity==8.5.0
python-docx==1.1.2
fpdf2==2.7.9
reportlab==4.4.3
supabase==2.18.1
supabase-auth==2.12.3
supabase-functions==0.10.1
storage3==0.12.1
realtime==2.7.0
